% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FROG_nonpal.R
\name{FROG_nonpal}
\alias{FROG_nonpal}
\title{Generate a fragment distance vector for your digested DNAString object}
\usage{
FROG_nonpal(my_DNA, pattern1 = "ATGCTGATT", pattern2 = "CGTGACTGGA")
}
\arguments{
\item{my_DNA, }{That's your DNAString you aim to digest}

\item{pattern1, }{the non-PALINDROMIC sequence pattern the restriction enzyme 1 recognizes}

\item{pattern2, }{OPTIONAL: the non-PALINDROMIC nucleotide sequence pattern the restriction enzyme 2 recognizes}
}
\value{
a fragment length vector
}
\description{
Generate a fragment distance vector for your digested DNAString object
}
\examples{
my_DNA<-gen_DNA(1000)
pattern1<-'ATGCTGATT'
pattern2<-'CGTGACTGGA'
FROG_pal(my_DNA, pattern1, pattern2)

library(BSgenome.Hsapiens.UCSC.hg38)
chr1_Hsapiens<-BSgenome.Hsapiens.UCSC.hg38$chr1

FROG_pal(chr1_Hsapiens, 'ATGCTGATT')
}
