% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FROG_plot.R
\name{FROG_plot}
\alias{FROG_plot}
\title{Plot the saved vector (fragment length) outputs generated with FROG_pal, FROG_nonpal or FROG_mix}
\usage{
FROG_plot(x, pattern1 = "", pattern2 = "")
}
\arguments{
\item{x, }{Saved output from FROG_pal, FROG_nonpal or FROG_mix (alternatively, an integer vector)}

\item{pattern1, }{Pattern1 used for digesting the DNAString}

\item{pattern2, }{Pattern2 used for digesting the DNAString}
}
\value{
Fragment length distributions and a table with basic descriptive statistics
}
\description{
Plot the saved vector (fragment length) outputs generated with FROG_pal, FROG_nonpal or FROG_mix
}
\examples{
library(tidyverse)
library(ggplot2)
library(ggpubr)
my_DNA<-gen_DNA(1000)
pattern1<-'GATC'
FROG_pal_output<-FROG_pal(my_DNA, pattern1)
FROG_plot(FROG_pal_output, pattern1)
}
